Move image preprocessing static method to preprocessing module
